requirements/dev.txt
# Development Requirements
# Include base requirements
-r base.txt

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
factory-boy==3.3.0
faker==22.0.0
playwright==1.41.1

# Code Quality
black==23.12.1
isort==5.13.2
flake8==7.0.0
pylint==3.0.3
mypy==1.8.0
django-stubs==4.2.7

# Debugging
django-debug-toolbar==4.2.0
ipython==8.20.0
ipdb==0.13.13

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Security
safety==3.0.1
bandit==1.7.6

# Performance
django-silk==5.0.4
